requirements.txt
numpy
requests
scipy
Pillow
fastdtw
editdistance
wget
imageio
